Security and architecture improvements
Security
--------

The IV is no longer hard-coded.

The id="pix" element is no longer required. This removes the only
previous traffic indicator that snique was in use.

Any resource referrenced by src= can now carry a snique eTag, not just
images.

Any hidden message is only revealed in an initially-hidden status bar
notification. This improves security if you are only able to surf the
web in public but are able to retire to a private location to read your
message afterwards.

Architecture
------------

Decoding is now performed in SniqueMessageDecoder class. This decouples
the decoding loginc from the hex-acquisition logic making it simpler to
use sources other than HTTP eTags in the future.