New icons, message id for status alerts, parse content-type header
New icons for launcher and status bar make this look more like a proper app.

The first 4 bytes of the IV are used as the message id for the status alert. It
is therefore now possible to retrieve multiple snique messages in a session and
review them all later. This functionality has not been tested.

The HTTP content-type header is now parsed to retrieve the character set and
MIME type of the retrieved content. This should avert the previously-frequent
occurence of the webview containing garbage.