Improve androidiness, browseriness
The logic to look for a snique message in a sequence of bytes is now a Service.
This allows external applications to access the snique-decoding service as well
as hopefully making the transition to multiple tabbed browser windows simpler.

The title of the activity is now set to the title of the web page, making it
more browser-like.